[HdSt] Fixed GL includes and comments
Minor cleanups now that HdStBufferResource and
HdSt_Subdivision computations wrap HgiBuffer
resources.
.

(Internal change: 2155553)

+++ pxr/imaging/hdSt/subdivision.h
@@ -221,7 +221,7 @@ public:
         size_t GetNumElements() const {
             return HdGetComponentCount(_resource->GetTupleType().type);
         }
-        GLuint BindVBO() {
+        uint64_t BindVBO() {
             return _resource->GetId()->GetRawResource();
         }
         HdStBufferResourceSharedPtr _resource;
